chore: Rename publish-button to commands-publish-button
