+++ tests/renderer/components/commands-spec.tsx
@@ -15,7 +15,7 @@ jest.mock('../../../src/renderer/components/commands-address-bar', () => ({
   AddressBar: 'address-bar'
 }));
 
-jest.mock('../../../src/renderer/components/publish-button', () => ({
+jest.mock('../../../src/renderer/components/commands-publish-button', () => ({
   PublishButton: 'publish-button'
 }));
